requirements.txt
Flask==3.0.0
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
reportlab==4.0.7
matplotlib==3.8.2
pandas==2.1.3
python-dotenv==1.0.0
